fix(database.db ConvertidorUsuario.java FrmMenuDinamico.java Persona.java usuario.java) I changed DNIoRUC Attribute to long and every method or class related to the use of that attribute

+++ src/ServiciosUsuario/ConvertidorUsuario.java
@@ -20,10 +20,10 @@ public class ConvertidorUsuario {
         return VerificadorUsuario.esTelefonoValido(resultado)?resultado:null;
     }
     
-    public static Integer tryParseUsuarioDNI(String usuarioDNI){
-        int resultado;
+    public static Long tryParseUsuarioDNI(String usuarioDNI){
+        long resultado;
         try{
-        resultado=Integer.parseInt(usuarioDNI);
+        resultado=Long.parseLong(usuarioDNI);
         }catch(NumberFormatException e){
             return null;
         }
